// src/lib/generated/OperationLineView.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ItemId } from "./ItemId";
import type { LocationId } from "./LocationId";
import type { OperationId } from "./OperationId";
import type { Quantity } from "./Quantity";

export type OperationLineView = { id: number, operationId: OperationId, kind: string, performedAt: string, performedBy: string | null, note: string | null, itemId: ItemId, itemName: string, fromLocationId: LocationId | null, fromLocation: string | null, toLocationId: LocationId | null, toLocation: string | null, quantity: Quantity, unitPrice: number | null, };